Fixed issue #2. Annoying one.

+++ src/io.f90
@@ -22,7 +22,7 @@ subroutine print_rdf(ihoo,ihoh,ihhh,na,boxlxyz,ng,nb,nrdf)
     ! ------------------------------------------------------------------
     ! Outputs RDF
     ! ------------------------------------------------------------------
-    integer(8) ihoo(imaxbin),ihoh(imaxbin),ihhh(imaxbin)
+    real(8) ihoo(imaxbin),ihoh(imaxbin),ihhh(imaxbin)
     integer na,ng,nb,nrdf
     integer no,nh,i
     real(8) delr,rlower,rupper,rh,vol,boxlxyz(3)
